requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.9.1
asteval==1.0.6
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
boto3==1.40.12
botocore==1.40.12
bw-migrations==0.2
bw2calc==2.1
bw2data==4.5
bw2io==0.9.11
bw2parameters==1.1.0
bw_processing==1.0
bw_simapro_csv==0.4.2
cached-property==2.0.1
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
constructive_geometries==1.0
country_converter==1.3.1
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
Deprecated==1.2.18
Django==5.2.5
django-unfold==0.64.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fsspec==2025.7.0
ftfy==6.3.1
greenlet==3.2.4
idna==3.10
ijson==3.4.0
inflection==0.5.1
isodate==0.7.2
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
linear-tsv==1.1.0
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
matrix_utils==0.6.2
morefs==0.2.2
mrio-common-metadata==0.2.1
multifunctional==1.1.1
numba==0.61.2
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
Pint==0.25
platformdirs==4.3.8
prettytable==3.16.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyecospold==4.0.0
PyJWT==2.10.1
pyparsing==3.2.3
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.2
randonneur==0.6.2
randonneur_data==0.6
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
requests==2.32.5
rfc3986==2.0.0
rpds-py==0.27.0
s3transfer==0.13.1
scipy==1.16.1
six==1.17.0
snowflake-id==1.0.2
SPARQLWrapper==2.0.0
sparse==0.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
stats_arrays==0.7
structlog==25.4.0
tableschema==1.21.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unfold==1.2.3
unicodecsv==0.14.1
uritemplate==4.2.0
urllib3==2.5.0
voluptuous==0.15.2
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.3
wurst==0.4
xlrd==2.0.2
xlsxwriter==3.2.5
